+++ sync-for-payables/src/codatsyncpayables/sdkconfiguration.py
@@ -20,9 +20,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '3.0.0'
-    sdk_version: str = '1.4.0'
-    gen_version: str = '2.139.1'
-    user_agent: str = 'speakeasy-sdk/python 1.4.0 2.139.1 3.0.0 codat-sync-for-payables'
+    sdk_version: str = '1.5.0'
+    gen_version: str = '2.143.2'
+    user_agent: str = 'speakeasy-sdk/python 1.5.0 2.143.2 3.0.0 codat-sync-for-payables'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> tuple[str, dict[str, str]]: